pyproject.toml
[tool.poetry]
name = "genai-python"
version = "0.1.0"
description = ""
authors = ["Your Name <illahibuxjagirani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.3.12"
langchain-community = "^0.3.3"
langchain-openai = "^0.2.3"
python-dotenv = "^1.0.1"
langchain = "^0.3.4"
huggingface-hub = "^0.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
